extract CORE article metadata
There is a new set of datastructures to represent the CORE article metadata.
I need to add some tests for this.
The other files have been renamed and moved about a little to make it a touch more obvious; there may be more refactoring yet.

The cmd has been updated to allow the fetching of a single metadata article.